requirements.txt
beautifulsoup4==4.9.0
bottle==0.12.18
bottle-websocket==0.2.9
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
dominate==2.5.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.3
future==0.18.2
gevent==1.5.0
gevent-websocket==0.10.1
greenlet==0.4.15
idna==2.9
IMDbPY==6.8
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.0
MarkupSafe==1.1.1
numpy==1.18.3
plotly==4.6.0
pycparser==2.20
requests==2.23.0
retrying==1.3.3
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.16
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
whichcraft==0.6.1
WTForms==2.2.1
